Introduce a CChainParameters singleton class.
Move hashGenesisBlock and pchMessageStart into this class and replace
references with calls to Params().MessageStart() and friends. The goal
is to reach a state where all numbers that vary between network
instances are collected in one place, then introduce a third set of
regtest parameters designed for the bitcoind/bitcoinj comparison tool
and app development.